# docs/openapi/spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API de Usuários",
    "version": "1.0.0",
    "description": "API REST de gestão de usuários e famílias, orientada a eventos."
  },
  "servers": [
    {
      "url": "http://localhost:8082",
      "description": "Local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/profile/family": {
      "post": {
        "tags": ["Family"],
        "summary": "Cria uma família",
        "description": "Cria uma nova família e adiciona o usuário autenticado como TITULAR. Publica o evento FAMILY_CREATED.",
        "operationId": "createFamily",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateFamilyRequest" },
              "example": { "nome": "Família Silva" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Família criada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FamilyResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "get": {
        "tags": ["Family"],
        "summary": "Obtém a família do usuário autenticado",
        "operationId": "getFamily",
        "responses": {
          "200": {
            "description": "Família retornada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FamilyResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/profile/family/members/{id}": {
      "put": {
        "tags": ["Family"],
        "summary": "Atualiza dados de um membro da família",
        "description": "Atualiza parentesco e status de um membro. Somente o titular ou o próprio membro podem alterar. Publica FAMILY_MEMBER_UPDATED.",
        "operationId": "updateFamilyMember",
        "parameters": [
          { "$ref": "#/components/parameters/memberId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FamilyUpdateMemberRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Membro atualizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FamilyMemberResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Family"],
        "summary": "Remove um membro da família",
        "description": "Somente o TITULAR pode remover membros. Publica FAMILY_MEMBER_REMOVED.",
        "operationId": "removeFamilyMember",
        "parameters": [
          { "$ref": "#/components/parameters/memberId" }
        ],
        "responses": {
          "204": { "description": "Membro removido" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/profile/family/invitations": {
      "post": {
        "tags": ["Family"],
        "summary": "Solicita um convite",
        "description": "Qualquer membro pode solicitar. Status inicial: AGUARDANDO_TITULAR. Publica FAMILY_INVITATION_REQUESTED.",
        "operationId": "requestInvitation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FamilyInvitationSendRequest" },
              "example": {
                "receiverEmail": "maria@email.com",
                "parentesco": "CONJUGE"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Convite solicitado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FamilyInvitationResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/profile/family/invitations/received": {
      "get": {
        "tags": ["Family"],
        "summary": "Lista convites recebidos",
        "description": "Lista convites com status PENDENTE recebidos pelo e-mail do usuário autenticado.",
        "operationId": "listReceivedInvitations",
        "responses": {
          "200": {
            "description": "Lista de convites recebidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/FamilyInvitationResponse" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/profile/family/invitations/sent": {
      "get": {
        "tags": ["Family"],
        "summary": "Lista convites enviados",
        "description": "Lista todos os convites enviados pela família do usuário autenticado.",
        "operationId": "listSentInvitations",
        "responses": {
          "200": {
            "description": "Lista de convites enviados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/FamilyInvitationResponse" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/profile/family/invitations/{id}/approve": {
      "put": {
        "tags": ["Family"],
        "summary": "Titular aprova convite",
        "description": "Muda status de AGUARDANDO_TITULAR para PENDENTE. Publica FAMILY_INVITATION_APPROVED_BY_TITULAR.",
        "operationId": "approveInvitation",
        "parameters": [
          { "$ref": "#/components/parameters/invitationId" }
        ],
        "responses": {
          "200": {
            "description": "Convite aprovado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FamilyInvitationResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/profile/family/invitations/{id}/reject-titular": {
      "put": {
        "tags": ["Family"],
        "summary": "Titular recusa convite",
        "description": "Titular rejeita o convite antes de enviar ao destinatário. Publica FAMILY_INVITATION_REJECTED_BY_TITULAR.",
        "operationId": "rejectInvitationByTitular",
        "parameters": [
          { "$ref": "#/components/parameters/invitationId" }
        ],
        "responses": {
          "200": {
            "description": "Convite recusado pelo titular",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FamilyInvitationResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/profile/family/invitations/{id}/accept": {
      "put": {
        "tags": ["Family"],
        "summary": "Destinatário aceita convite",
        "description": "Aceita um convite PENDENTE e adiciona o usuário à família. Publica FAMILY_INVITATION_ACCEPTED.",
        "operationId": "acceptInvitation",
        "parameters": [
          { "$ref": "#/components/parameters/invitationId" }
        ],
        "responses": {
          "200": {
            "description": "Convite aceito",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FamilyInvitationResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/profile/family/invitations/{id}/reject": {
      "put": {
        "tags": ["Family"],
        "summary": "Destinatário recusa convite",
        "description": "Destinatário rejeita um convite PENDENTE. Publica FAMILY_INVITATION_REJECTED.",
        "operationId": "rejectInvitation",
        "parameters": [
          { "$ref": "#/components/parameters/invitationId" }
        ],
        "responses": {
          "200": {
            "description": "Convite recusado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FamilyInvitationResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via Keycloak (dev-client)."
      }
    },
    "parameters": {
      "memberId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID do membro da família",
        "schema": { "type": "integer", "format": "int64", "example": 1 }
      },
      "invitationId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID do convite",
        "schema": { "type": "integer", "format": "int64", "example": 1 }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Token ausente ou inválido",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": 401, "error": "Unauthorized", "message": "Token ausente ou inválido" }
          }
        }
      },
      "Forbidden": {
        "description": "Acesso negado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": 403, "error": "Forbidden", "message": "Acesso negado" }
          }
        }
      },
      "NotFound": {
        "description": "Recurso não encontrado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": 404, "error": "Not Found", "message": "Recurso não encontrado" }
          }
        }
      },
      "Conflict": {
        "description": "Conflito de estado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": 409, "error": "Conflict", "message": "Usuário já pertence a uma família" }
          }
        }
      },
...
    },
    "schemas": {
...
      "CreateFamilyRequest": {
        "type": "object",
        "required": ["nome"],
        "properties": {
          "nome": { "type": "string", "description": "Nome da família", "example": "Família Silva" }
        }
      },
      "FamilyResponse": {
        "type": "object",
        "properties": {
          "id":      { "type": "integer", "format": "int64", "description": "Identificador único da família", "example": 1 },
          "nome":    { "type": "string", "description": "Nome da família", "example": "Família Silva" },
          "titular": { "$ref": "#/components/schemas/FamilyMemberResponse", "description": "Membro titular da família" },
          "membros": {
            "type": "array",
            "description": "Lista de membros da família, excluindo o titular",
            "items": { "$ref": "#/components/schemas/FamilyMemberResponse" }
          }
        }
      },
      "FamilyMemberResponse": {
        "type": "object",
        "properties": {
          "id":           { "type": "integer", "format": "int64", "description": "Identificador único do membro", "example": 1 },
          "nomeCompleto": { "type": "string", "description": "Nome completo do membro", "example": "João Silva" },
          "email":        { "type": "string", "format": "email", "description": "E-mail do membro", "example": "joao@email.com" },
          "parentesco":   { "allOf": [{ "$ref": "#/components/schemas/ParentescoEnum" }], "description": "Grau de parentesco do membro em relação ao titular" },
          "status":       { "allOf": [{ "$ref": "#/components/schemas/FamilyMemberStatusEnum" }], "description": "Situação atual do membro na família" }
        }
      },
      "FamilyUpdateMemberRequest": {
        "type": "object",
        "properties": {
          "parentesco": { "allOf": [{ "$ref": "#/components/schemas/ParentescoEnum" }], "description": "Novo grau de parentesco do membro" },
          "status":     { "allOf": [{ "$ref": "#/components/schemas/FamilyMemberStatusEnum" }], "description": "Novo status do membro" }
        }
      },
      "FamilyInvitationSendRequest": {
        "type": "object",
        "required": ["receiverEmail", "parentesco"],
        "properties": {
          "receiverEmail": { "type": "string", "format": "email", "description": "E-mail do destinatário do convite", "example": "maria@email.com" },
          "parentesco":    { "allOf": [{ "$ref": "#/components/schemas/ParentescoEnum" }], "description": "Grau de parentesco do convidado em relação ao titular" }
        }
      },
      "FamilyInvitationResponse": {
        "type": "object",
        "properties": {
          "id":            { "type": "integer", "format": "int64", "description": "Identificador único do convite", "example": 1 },
          "requesterNome": { "type": "string", "description": "Nome do membro que solicitou o convite", "example": "João Silva" },
          "receiverEmail": { "type": "string", "format": "email", "description": "E-mail do destinatário do convite", "example": "maria@email.com" },
          "parentesco":    { "allOf": [{ "$ref": "#/components/schemas/ParentescoEnum" }], "description": "Grau de parentesco do convidado" },
          "status":        { "allOf": [{ "$ref": "#/components/schemas/InvitationStatusEnum" }], "description": "Status atual do convite" },
          "createdAt":     { "type": "string", "format": "date-time", "description": "Data e hora de criação do convite em UTC", "example": "2026-05-03T19:00:00Z" }
        }
      },
      "ParentescoEnum": {
        "type": "string",
        "description": "Grau de parentesco do membro em relação ao titular da família",
        "enum": ["TITULAR", "CONJUGE", "FILHO", "FILHA", "PAI", "MAE", "IRMAO", "IRMA", "OUTRO"],
        "example": "CONJUGE"
      },
      "FamilyMemberStatusEnum": {
        "type": "string",
        "description": "Situação do membro na família",
        "enum": ["ATIVO", "INATIVO"],
        "example": "ATIVO"
      },
      "InvitationStatusEnum": {
        "type": "string",
        "description": "Status do fluxo de convite: AGUARDANDO_TITULAR (aguarda aprovação do titular), PENDENTE (aguarda aceite do destinatário), ACEITO, RECUSADO",
        "enum": ["AGUARDANDO_TITULAR", "PENDENTE", "ACEITO", "RECUSADO"],
        "example": "PENDENTE"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time", "description": "Data e hora do erro em UTC" },
          "status":    { "type": "integer", "description": "Código HTTP do erro", "example": 400 },
          "error":     { "type": "string", "description": "Descrição resumida do erro HTTP", "example": "Bad Request" },
          "message":   { "type": "string", "description": "Mensagem detalhada do erro", "example": "Mensagem de erro" },
          "path":      { "type": "string", "description": "Caminho do endpoint que gerou o erro", "example": "/api/v1/profile" }
        }
      }
    }
  }
}